+++ common/cpp/include/flutter_data_channel.h
@@ -10,6 +10,7 @@ class FlutterRTCDataChannelObserver : public RTCDataChannelObserver {
  public:
   FlutterRTCDataChannelObserver(scoped_refptr<RTCDataChannel> data_channel,
                                 BinaryMessenger* messenger,
+                                TaskRunner* task_runner,
                                 const std::string& channel_name);
   virtual ~FlutterRTCDataChannelObserver();
 
@@ -38,6 +39,9 @@ class FlutterDataChannel {
                        const std::string& type,
                        const EncodableValue& data,
                        std::unique_ptr<MethodResultProxy>);
+
+  void DataChannelGetBufferedAmount(RTCDataChannel* data_channel,
+                       std::unique_ptr<MethodResultProxy> result);
 
   void DataChannelClose(RTCDataChannel* data_channel,
                         const std::string& data_channel_uuid,