+++ lib/src/web/rtc_video_renderer_impl.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 import 'dart:js_interop';
-import 'dart:js_util' as jsutil;
 import 'dart:ui_web' as web_ui;
 
 import 'package:flutter/foundation.dart';
@@ -73,6 +72,9 @@ class RTCVideoRenderer extends ValueNotifier<RTCVideoValue>
   int get videoHeight => value.height.toInt();
 
   @override
+  RTCVideoValue get videoValue => value;
+
+  @override
   int get textureId => _textureId;
 
   @override
@@ -90,12 +92,12 @@ class RTCVideoRenderer extends ValueNotifier<RTCVideoValue>
 
   String get viewType => 'RTCVideoRenderer-$textureId';
 
-  void _updateAllValues() {
-    final element = findHtmlView();
+  void _updateAllValues(web.HTMLVideoElement fallback) {
+    final element = findHtmlView() ?? fallback;
     value = value.copyWith(
       rotation: 0,
-      width: element?.videoWidth.toDouble() ?? 0.0,
-      height: element?.videoHeight.toDouble() ?? 0.0,
+      width: element.videoWidth.toDouble(),
+      height: element.videoHeight.toDouble(),
       renderVideo: renderVideo,
     );
   }
@@ -241,10 +243,8 @@ class RTCVideoRenderer extends ValueNotifier<RTCVideoValue>
   Future<bool> audioOutput(String deviceId) async {
     try {
       final element = _audioElement;
-      if (null != element && jsutil.hasProperty(element, 'setSinkId')) {
-        await jsutil.promiseToFuture<void>(
-            jsutil.callMethod(element, 'setSinkId', [deviceId]));
-
+      if (null != element) {
+        await element.setSinkId(deviceId).toDart;
         return true;
       }
     } catch (e) {
@@ -273,13 +273,13 @@ class RTCVideoRenderer extends ValueNotifier<RTCVideoValue>
 
       _subscriptions.add(
         element.onCanPlay.listen((dynamic _) {
-          _updateAllValues();
+          _updateAllValues(element);
         }),
       );
 
       _subscriptions.add(
         element.onResize.listen((dynamic _) {
-          _updateAllValues();
+          _updateAllValues(element);
           onResize?.call();
         }),
       );